feat(npm-scripts): add getTypeScriptBuildOrder() plus helper
Given a dependency graph of TypeScript projects,
`getTypeScriptBuildOrder()` returns a suitable build-order via
topological traversal (reporting cycles if any are found). This one
lends itself nicely to being unit tested, so added some tests for it.

To actually be useful though, you need the dependency graph, so I added
`getTypeScriptDependencyGraph()` for that. That one looks at the
file-system and is a bit more cumbersome to test from unit tests; also,
it's pretty straightforward, so I just verified that it works manually
and that is probably going to be enough.

Test plan:

- Tested `getTypeScriptBuildOrder()` by running the unit tests.
- Tested `getTypeScriptDependencyGraph()` by copying over to
  `liferay-portal` checkout and running with some `console.log()`.