refactor(npm-scripts): extract color() function for reuse
I am going to want colorized output in the next commit, so putting this
somewhere where I can reuse it.

You may be wondering why we're not using a third-party dependency like
`chalk` (https://www.npmjs.com/package/chalk) which had an impressive 84
million downloads in the last week. The short version is that we like
keeping our dependency footprint small and this 15-line function is
totally adequate for our needs. chalk, on the other hand, at the time of
writing:

- Is a repo with 311 commits and 32 files on its `main` branch
  (excluding dependencies).
- Takes 13 seconds to `npm install --production` (skipping its
  `devDependencies`).
- Has two vulnerabilities in its production dependency graph (which
  consists of `ansi-styles`, `color-convert`, `color-name`,
  `supports-color` and `has-flag`), as reported by `npm audit`.
- Has production dependencies occupy 136KB on disk.