feat(npm-scripts): add "types" subcommand
As the README says, this subcommand freshens all the ".d.ts" and
"tsconfig.tsbuildinfo" files in the repo by running `tsc` on the
projects in dependency order.

I initially thought of calling this subcommand "typegen", but looking at
our other command names, I saw they are almost all verbs, and all single
words. The ones that aren't verbs like "build", "test", "check" etc are
generally named after the related tool (eg. "webpack", "prettier",
"storybook") or concept (eg. "theme"). So, "types" wins, because it is a
single word, and refers to the concept that it is responsible for
managing.

This is a global process, so if you run it from inside a project, you'll
see something like this:

    You ran "liferay-npm-scripts types" from:

        apps/frontend-js/frontend-js-react-web

    But generating types is a global process; will run from:

        /Users/greghurrell/code/portal/liferay-portal/modules

    Generating types (1 of 3): @liferay/frontend-js-state-web

    ***************
    *** WARNING ***
    ***************

    Out-of-date TypeScript build artifacts at:

        apps/frontend-js/frontend-js-state-web/types

    Please commit them.

    M modules/apps/frontend-js/frontend-js-state-web/types/tsconfig.tsbuildinfo

    Generating types (2 of 3): @liferay/frontend-js-react-web

    ***************
    *** WARNING ***
    ***************

    Out-of-date TypeScript build artifacts at:

        apps/frontend-js/frontend-js-react-web/types

    Please commit them.

    M modules/apps/frontend-js/frontend-js-react-web/types/tsconfig.tsbuildinfo

    Generating types (3 of 3): remote-app-client-js

Note that I only see those warnings because my local state _is_ out of
date. If you're up-to-date, you'll just see:

    Generating types (1 of 3): @liferay/frontend-js-state-web
    Generating types (2 of 3): @liferay/frontend-js-react-web
    Generating types (3 of 3): remote-app-client-js

Obviously, if you run this from the top-level "modules/" directory, you
don't see the "generating types is a global process" message.

On my machine, this takes about 5 seconds. In a subsequent commit I am
going to extend this as a mechanism for checking that type information
is up-to-date in CI. It's probably fast enough to run as part of
`ci:test:sf` on any changes that touch projects with TypeScript in them;
this is stretching the definition of "formatting", but Liferay let that
ship sail a long time ago.

(Minor clarification: when I say "let that ship sail a long time ago", I
actually mean, "designed, developed, and willingly strapped itself onto
that intercontinental ballistic missile a long time ago").

Test plan: Ran the commands as described above.