feat(npm-scripts): teach "types" to report staleness issues
As mentioned in the previous commit, we can use this from CI to detect
staleness issues and report them as part of `ci:test:sf`.

If you run it with stale artifacts you are going to see something like:

    Generating types (1 of 3): @liferay/frontend-js-state-web

    ***************
    *** WARNING ***
    ***************

    Out-of-date TypeScript build artifacts at:

        apps/frontend-js/frontend-js-state-web/types

    Please commit them.

    M modules/apps/frontend-js/frontend-js-state-web/types/tsconfig.tsbuildinfo

    Generating types (2 of 3): @liferay/frontend-js-react-web

    ***************
    *** WARNING ***
    ***************

    Out-of-date TypeScript build artifacts at:

        apps/frontend-js/frontend-js-react-web/types

    Please commit them.

    M modules/apps/frontend-js/frontend-js-react-web/types/tsconfig.tsbuildinfo

    Generating types (3 of 3): remote-app-client-js
    Error: Type generation was successful but 2 projects were out of date. Please commit updated versions of the artifacts in the projects listed above.

And the command will exit with a non-zero status. If everything is
up-to-date, you get a zero status, and the output looks like this:

    Generating types (1 of 3): @liferay/frontend-js-state-web
    Generating types (2 of 3): @liferay/frontend-js-react-web
    Generating types (3 of 3): remote-app-client-js

Next commit will complete the wiring-up here by baking this into our
pre-flight checks on any commits that touch projects with TypeScript
files in them.