feat(npm-scripts): check for TS artifact staleness during CI runs
Basically does what it says on the tin. If you have touched any paths
inside a TS project then we'll run our "types" subcommand to verify that
the committed TS artifacts are up-to-date.

I don't know whether "preflight" is a great _name_ for this phase but it
does seem to be the right _place_. Currently when you run "check" (which
is what CI runs) we will run:

1.  "preflight" checks; which consist of:

    a)  checking for banned config file names

    b)  checking for banned dependencies inside package.json files (note
        that we may and probably will add more package.json checks in
        the future)

    c)  checking that TS artifacts are up to date

2.  "format" (Prettier) checks

3.  "lint" (ESLint) checks

In other words, "preflight" means "everything that isn't handled by
Prettier and ESLint".

So, as noted in the commit, this will only have effect in CI, because
that's the only place where `LIFERAY_NPM_SCRIPTS_WORKING_BRANCH_NAME`
will be defined. At least for now, this check is plenty fast (only a few
seconds). We can worry about performance later on, and we probably have
a fair bit of time before that happens because TS adoption will likely
be quite a long and drawn-out affair.

Note that we want this to work on in CI in order to keep the local
development experience snappy. If you are working in
"frontend-js-react-web", you shouldn't have to wait for a global check
to run just because you ran "yarn run check"; instead, you should be
able to rely on CI to keep `master` in a good state, and know that all
the other projects are up-to-date and that you only need worry about
your own.

Test plan:

Copy this over to liferay-portal and see `yarn run check` fail like this
when you set `LIFERAY_NPM_SCRIPTS_WORKING_BRANCH_NAME` to `master` and
have stale artifacts. If the artifacts aren't stale, or you don't have
that environment variable, you don't see an error.

    Preflight check failed:
    checkTypeScriptTypeArtifacts() failed: Error: Type generation was successful but 2 projects were out of date. Please commit updated versions of the artifacts in the projects listed above.