test(npm-scripts): add small pause to ensure tests pass on Linux
Yes, this is unrelated to the neighboring commits, but today I'm working
on Linux and I am seeing intermittent test failures in this file:

    ● sass › rebuilds sass if any file has been modified

      expect(received).toBeGreaterThan(expected)

      Expected: > 1617616133385
      Received:   1617616133385

At first I thought the machine just may be too fast so it sometimes
fails, so I switched to millisecond-granularity timestamps. But that
didn't work:

    ● sass › rebuilds sass if any file has been modified

      expect(received).toBeGreaterThan(expected)

      Expected: > 1617616326854.6343
      Received:   1617616326854.6343

So, evidently the filesystem here is doing some kind of lazy flush which
means that we don't necessarily see an update.

The "fix" is to add a small pause (100ms) in a few places. I originally
added a 1 second pause in each place, but experimentation showed that
even the smaller pause is enough to make the tests pass reliably (I ran
them about 10 times with this interval).

This does slow the tests down, obviously, by 0.4 seconds total, but I think we can
live with that, as it is a small part of the overall run; for example,
last CI run on `master` took 3m14s, so this change would have grown that
to 3m14.4s, an increase of 0.2%:

    https://github.com/liferay/liferay-frontend-projects/runs/2218144611

The other possible solution here would be to rewrite the test in such a
way that it relied on mocks instead of filesystem inspection, but that
might just be trading one smell for another, so going with this for now.