feat(npm-scripts): configure TypeScript as part of `runTSC()`
I originally thought I would have to do this in two places in order to
have it take effect in both "build" and "check", but it turns out that
one place (`runTSC()`) is enough, because it gets called in both cases.

- If you run "build", we invoke `runTSC()` and that automatically sets
  up (or freshens) the TypeScript config for you.
- If you run "check", we invoke the "types" subcommand, and that one
  ends up using `runTSC()` under the covers, so we once again freshen
  the config for you.

When running "build" you'll get a big fat warning in the console
reminding you to commit the changed config (if there are changes,
otherwise we print nothing).

When running "check", if any of the involved projects had changes, those
will bubble up and cause `ci:test:sf` to fail, which is what we want.

Test plan: Copy all this over to liferay-portal checkout and mutate a
tsconfig.json. See `yarn check` fail when you have
`LIFERAY_NPM_SCRIPTS_WORKING_BRANCH_NAME` set, otherwise not. Likewise,
observe how a `yarn build` updates the artifacts and prints a warning
when they are not up-to-date.