fix(npm-scripts): fix test failures on Windows
As predicted here:

    https://github.com/liferay/liferay-frontend-projects/pull/478#issuecomment-811274916

we need to tweak our paths to keep the tests passing on Windows.

    https://github.com/liferay/liferay-frontend-projects/runs/2269174696

It's quite possible that everything works in practice on Windows even in
the presence of these test failures, seeing as it will for the most part
accept either "/" or "\" as a separator, but whatever...

My initial attempt of using `path.posix` in various places didn't work.
We wanted:

    "../frontend-js-state-web/src/main/resources/META-INF/resources/index.ts"

and were getting:

    "..\\frontend-js-state-web\\src\\main\\resources\\META-INF\\resources\\index.ts"

but after the "fix" we got:

    "D:\\a\\liferay-frontend-projects\\liferay-frontend-projects\\projects\\npm-tools\\packages\\npm-scripts\\__fixtures__\\typescript\\modules/apps/frontend-js/frontend-js-state-web/src/main/resources/META-INF/resources/index.ts"

as well as a slew of other failures:

    https://github.com/liferay/liferay-frontend-projects/runs/2269241475

So, let's go for this instead.