fix(npm-scripts): add missing `--jsx react` configuration
Discovered while testing:

    https://github.com/liferay/liferay-frontend-projects/releases/tag/npm-scripts%2Fv42.0.0-pre.0

I had this in my cheatsheet, but neglected to include it in the base
config:

    https://gist.github.com/wincent/42bc9fa924770df92ae876efbbe4d634

One sneaky thing to note: even after adding this config, you still might
get:

    TS17004: Cannot use JSX unless the '--jsx' flag is provided.

Because of caching. Blowing about the `types/` directories fixes it.

Note that there are some comments in here about tsconfig.json files
getting cached and not-reread:

    https://stackoverflow.com/a/64969461/2103996