refactor(npm-scripts): store tsbuildinfo in tmp/
Because, to my horror, after testing this on Linux, I have discovered
two things which make it machine specific.

First, it stores absolute paths under `pathsBasePath` keys:

-      "pathsBasePath": "/Users/greghurrell/code/portal/liferay-portal/modules/apps/frontend-js/frontend-js-react-web",
+      "pathsBasePath": "/home/glh/code/portal/liferay-portal/modules/apps/frontend-js/frontend-js-react-web",

Second, on macOS (with its wonderful case-insensitive APFS filesystem),
we see the paths all "normalized" to lowercase:

-      "../src/main/resources/meta-inf/resources/state.ts": {
+      "../src/main/resources/META-INF/resources/State.ts": {

Now, I found that if I generated the file on Linux and copied it over to
macOS, at least macOS didn't set the case back again, but that doesn't
help us much (ie. we don't want to have only people on Linux producing
valid mappings).

In any case, the absolute paths are a deal-breaker. So, let's store the
files under "tmp/" instead (`.gitignored` already) so that they don't
get committed. What these will do is speed up the subsequent rebuilds
because `composite` implies `incremental`:

- https://www.typescriptlang.org/tsconfig#composite
- https://www.typescriptlang.org/tsconfig#incremental
- https://www.typescriptlang.org/tsconfig#tsBuildInfoFile

Note that it's fine for this file to be omitted from the repo as it is
not essential to the operation of the build:

> These `.tsbuildinfo` files can be safely deleted and don’t have any
> impact on our code at runtime - they’re purely used to make compilations
> faster. We can also name them anything that we want, and place them
> anywhere we want using the `--tsBuildInfoFile` flag.

https://www.typescriptlang.org/docs/handbook/release-notes/typescript-3-4.html#faster-subsequent-builds-with-the---incremental-flag

So, if I have this right, if you pull down the repo and do an initial
`ant all`, any subesquent build will be faster (although the usual
caveats about `clean`-ing and deleting `tmp` apply).