style(npm-scripts): add SF-friendly `stringify()` function
To stop SF failures like this one:

    https://github.com/liferay-frontend/liferay-portal/pull/942#issuecomment-813498234

SF wants to format empty arrays like this:

    [
    ]

and empty objects like this:

    {
    }

Instead of like `JSON.stringify()`, which formats them like:

    []

and:

    {}

respectively.

So, "simply" write a 50 line function to do this and 200 lines of tests.
Note that it assumes well-formed input (no circular references) and
doesn't try to do any special handling of exotic object types.

Writing this code might seem like insanity, but we discussed on previous
PRs the possibility of letting Prettier handle formatting of JSON in
liferay-portal, but decided against it:

    https://github.com/liferay-frontend/liferay-portal/pull/871#issuecomment-791558324

Test plan: run unit tests.