fix(npm-scripts): don't post-process TypeScript if there is no TS
Noticed while preparing:

    https://github.com/liferay-frontend/liferay-portal/pull/970

yesterday that if you try to start porting a module to TypeScript by
doing this:

    touch tsconfig.json
    yarn run build

it will die with:

    ENOENT: no such file or directory, scandir './types'

This happens because you haven't created any `.ts` files yet, and so
`tsc` never creates the `types` directory. That in turn means that when
we try to post-process everything matching `**/*.d.ts` in `types/` glob
expansion is going to complain that there _is_ no `types/` directory.

Test plan:

    # Before this commit, in liferay-portal project (eg. wiki-web):

    touch tsconfig.json
    yarn run build # see ENOENT

    # After copying this commit:

    touch tsconfig.json
    yarn run build # see tsconfig.json get filled out; no error