refactor(npm-scripts): split preflight checks into separate files
Right now we have exactly three "preflight" checks, and they all do
exactly one thing. In the name of simplicity, we kept them in one file.

But after this commit, we're going to be adding a new check in service
of:

    https://github.com/liferay/liferay-frontend-projects/issues/469

which will mean that our `checkPackageJSONFiles` function is going to
have two responsibilities now.

In order to add this new kind of check without muddying the waters, we
are splitting the checks into separate files in this commit. We always
knew that this day would come, which is why we made the `check`
directory in the first place (ie. because we knew we'd eventually have a
bunch of "stuff" to put in it).