feat(npm-scripts)!: enforce use of `@liferay/` named scope
Ideally we'd have renamed all the packages (as part of
https://issues.liferay.com/browse/LPS-129670), but that is going to be a
huge task that we are not likely to get to in a hurry (especially
because in order to be complete we'd have to do it on multiple
branches). So in this commit, we're taking the "allow list" approach,
which means banning every package name that isn't in the named scope
_except_ some pre-set list of exceptions containing some 200 odd names
that currently exist in our repos/branches and that we own and therefore
consider to be relatively safe with respect to dependency confusion
attacks:

    https://github.com/liferay/liferay-frontend-projects/blob/314d612abaea193d26b0f4c496e65bc5228c2882/guidelines/general/security/dependency_confusion.md

The names come from this spreadsheet:

    https://docs.google.com/spreadsheets/d/1jXsn5ldB6IehQTsnggWHJKNdo-1m1Mmuoc26lh12saU/edit

which we prepared a couple of months ago, so it may be out of date. My
test plan should catch any omissions that have snuck into liferay-portal
in the meantime.

Test plan: Copy files over to liferay-portal repo and run `yarn
checkFormat` from the top. See no errors reported. Edit a package name
and retry. See errors reported.

Closes: https://github.com/liferay/liferay-frontend-projects/issues/469