fix(npm-scripts): don't complain about packages without a name
In testing this PR:

    https://github.com/liferay/liferay-frontend-projects/pull/500

Saw this:

    package.json: BAD - error thrown during checks: TypeError: Cannot read property 'startsWith' of undefined

So, I thought, let's make that message a bit better:

    package.json: BAD - package package.json does not have a name

But after testing it, I realized that a package is pretty useless
without a name, and you'll normally notice it if you don't have one (ie.
the bundler won't work, so it's unlikely this kind of thing would get
into `master`).

However, there is one place in liferay-portal where we legitimately
don't have any need for a name and that is right here:

    https://github.com/liferay/liferay-portal/blob/master/modules/package.json

So, let's just graciously skip over that case.

Test plan: Copy this over to a liferay-portal checkout and run `yarn
checkFormat` and see it print no error for the top-level
`modules/package.json`.