refactor(senna)!: trim two dependencies off the dependency graph
Due to the aforementioned headaches trying to get the build to work (or
even `yarn`!) I am taking the opportunity to trim off a couple of
dependencies, just to reduce the surface area. These correspond to tasks
not in the "critical path" that we have to follow in order to produce a
working build.