build(senna): peg rollup-stream to v1.23.1
Prior to this change:

    Error: Unexpected key 'moduleName' found, expected one of: acorn,
    amd, banner, cache, context, entry, exports, extend, external, file,
    footer, format, globals, indent, input, interop, intro, legacy,
    moduleContext, name, noConflict, onwarn, output, outro, paths,
    plugins, preferConst, pureExternalModules, sourcemap, sourcemapFile,
    strict, targets, treeshake, watch

        at validateKeys (maintenance/projects/senna/node_modules/rollup/dist/rollup.js:172:11)

So peg the version at the one we had in the package-lock.json (v1.23.1).
I also tried pegging it to the bottom of the range as specified in the
package.json (v1.18.0), but in both cases we get this error:

    Error: 'default' is not exported by node_modules/metal-uri/lib/Uri.js

        at error (maintenance/projects/senna/node_modules/rollup/src/utils/error.js:9:14)

This is progress, so I am going to commit it and keep looking.