build(senna): depend on metal-uri 2.2.6 exactly
2.2.6 is the version we were using in the old package-lock.json file.

So, this is a bit sad, but with this change, the build works! Yay.

Why sad? Because setting "resolutions" alone didn't work; Yarn kept
putting in a higher version (2.4.0), even when I ran `npx
yarn-deduplicate yarn.lock`. I am concerned that this may not be great
when moving to liferay-portal's older branches (unless my `yarn why`
deceives me, it is `2.4.0` on each of 7.1.x, 7.2.x, and 7.3.x).

So, going to see if I can trick it into using 2.2.6 in the lockfile
without pinning it like I am in the `package.json` in the next commit.