docs(monorepo): provide example of how to "rebase" (redo) subtree merge
As this may be non-obvious, want to document what I did to update my
"merge Senna history" PR that had been sitting abandoned for three
months.

Seeing as we now have quite a few examples, and they are wordy, added
some headings to make it easier to know where you are in the doc.